backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==1.10.13
pytest==7.4.3
httpx==0.25.2
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1